Partition provider test by language
Use inverted Go build flags to not disable test by default.
